requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langchain-groq==0.1.9

openai>=1.30.0
python-dotenv
streamlit
scikit-learn
ipykernel

pypdf
pymupdf
beautifulsoup4
arxiv
wikipedia

faiss-cpu
sentence-transformers
pandas
duckdb
chromadb
numpy
